[CSL-1119] join Kademlia after node is up
For reasons described in the ticket, it would be very invasive to have
the Kademlia instance itself started from within the NodeAction
callback. The solution in this commit is to refrain from joining the
network until the node is up. It's possible that this node will be
contacted on its Kademlia UDP socket before the node is ready, but
that's unlikely. When it joins the network, broadcasting its own
presence, it is ready to accept application connections.