[CSL-1120] remove queryBlocksWorker
This worker should not be included in full nodes (they're addressable,
blocks will be pushed to them so no need to query).

This worker perhaps *should* be included in wallets, because they are
unaddressable and blocks cannot be pushed *unsolicited* to them.
However, this worker was not included as a worker in the wallet, and in
any case should not be included. A better approach is to subscribe to
new blocks and have the full nodes push them as they arrive. That's
TODO.