Try benchmarks with >= 20 nodes -- DO NOT MERGE
This disables block relaying, which probably contributed to stability
problems in the setup of the benchmarks for 20 nodes or more, where
we're trying to not use relaying, and instead communicate published
blocks to every node (by making each node every node's neighbour).  Any
kind of further relaying will lead to quadratic communication in this
setup, which has to be avoided.

Also, changed the k parameter to get epoch lengths of one hour.