[CSL-1617] genesis constants
Values which are determined by the genesis data are removed from the raw
core config. The module Pos.Core.Genesis exports them, still pretending
that they are global constants (they're not, see newInitFuture).

Some crypto (Pos.Crypto.SignTag) needs to use the genesis "global
constants"; the protocolMagic in particular. But the genesis module
depends transitively upon this, so we're stuck. The knot is broken with
an hs-boot. Using an hs-boot may be inevitable unless signTag takes the
protocol magic as a parameter: there has to be some module which defines
the "global constants", and it's definitely going to mention
domain-specific types from Pos.Core.Types, and those types often mention
crypto, and crypto mentions global constants. Tough. There may be a way
to get around it by factoring Pos.Core.Types more aggressively.

I noticed that "global constants" are used in typeclass instances
(Enum SlotOrEpoch in particular). This means we really must present
them as global constants. We could keep up the newInitFuture mechanism,
or use Data.Reflection instead, as we already do for the impotent
CoreConstants type in Pos.Core.Context.Const. We may want to extend that
type to include all CoreConfig and GenesisData, and deliver it via
Data.Reflection as we do now

  giveStaticConsts
   :: CoreConfig
   -> GenesisData
   -> (HasCoreConstants => r)
   -> r

Seems like exactly the use case Data.Reflection was intended for.