[CSL-1794] Update versions in configuration
We are going to propose and vote for an update to block version 0.1.0.
We will also need to set application version to 1 for cardano-sl.
Previosuly cardano-sl software was launched with version 2, but it was by
mistake, because last confirmed version was 0. This mistake didn't have any
impact though.
Wallet configs now don't reuse `mainnet_base_update`, but instead define
all versions explicitly, because the previous solution was error-prone.
Wallet configs effectively didn't change.
In `mainnet_full` config (for core nodes) two values changed:
`applicationVersion' and `bvMinor'.