[CSL-2103] Get rid of the MonadBalances abstraction
This PR removes the `MonadBalances` abstraction and paves the way to a
uniform way of accessing the database during HTTP handlers.

Code in the `auxx` tool has been patched in a best-effort manner, and
will need to be reviewed at a later stage to ensure it has preserved its
old behaviour.